[refactor] #추천코스 장원태 [231231] - courseadd.js의 유틸 및 상수 분리
