+++ src/main/webapp/resources/course/script/restriction.js
@@ -1,0 +1,8 @@
+export const MAX_SPOTS_IN_COURSE = 5; // 하나의 코스에 들어갈 수 있는 여행지 최대 개수
+
+// 각 input의 최대 길이
+export const INPUT_MAX_LENGTH = {
+		"name": 100,
+		"briefcontent": 500,
+		"longdetail": 2000
+};
